pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "watchtower"
version = "1.0.0"
authors = [
    { name="Your Name", email="your.email@example.com" },
]
description = "A desktop security application for motion and human detection using webcam"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "opencv-python>=4.5.0",
    "numpy>=1.19.0",
    "pillow>=8.0.0",
]

[project.scripts]
watchtower = "watchtower.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["watchtower*"]
namespaces = false 